+++ pages/api/confess/index.js
@@ -7,8 +7,8 @@ module.exports = async ({ body, headers: { token }}, res) => {
 		}catch(e){}
 	}
 	try {
-		const {confession, filename} = body;
-		const id = await addToQueue(confession, null, filename, user?.uid);
+		const {confession, filename, poll} = body;
+		const id = await addToQueue(confession, null, filename, user?.uid, { poll });
 		return res.json({id});
 	}catch(error){
 		console.error(error);