+++ utils/objectHelper.js
@@ -1,1 +1,4 @@
-export const removeEmpty = obj => Object.fromEntries(Object.entries(obj).filter(([, value]) => value !== undefined ))
+export const removeNullish = obj =>
+	Object.fromEntries(Object.entries(obj)
+			.filter(([, value]) => value != undefined
+	));